The check was necessary; the expensive overload was not
Bounding `NtsArrayL.get` by length fixed a silent stale-slot read and cost
`growth-grown` 26% -- 1.01x to 1.28x. Both in one commit, and only the first was
argued for.

`get` takes a `double`, so it was calling `bounds(int, double)`, which re-proves
*integrality* with a `(double)(int)` round trip per element. This subscript is
emitted `checked: false`: the middle end already proved the index whole and only
its range was ever in question. Casting once and calling the `(int, int)`
overload keeps the range check and drops the proof of a precondition.

    growth-grown   1.01x -> 1.28x -> 1.02x

So the correctness fix costs about one percent, not twenty-six. Nothing in the
argument for *whether* to check said anything about *which* check, and I picked
the one whose signature happened to match the caller.

`growth-fixed` is unmoved at 1.00x, which is the control: it holds a `double[]`
and never enters the growable path.

+++ runtime/jvm/src/nts/rt/NtsArrayL.java
@@ -32,7 +32,13 @@ public final class NtsArrayL {
      * Bounding by {@code a.length} answers both.
      */
     public static Object get(NtsArrayL a, double at) {
-        return a.items[NtsRuntime.bounds(a.length, at)];
+        // The `(int, int)` overload, not the `(int, double)` one: this
+        // subscript is emitted `checked: false`, so the middle end already
+        // proved the index integral and only its *range* is in question.
+        // The double form re-proves integrality with a `(double)(int)`
+        // round trip per element, which cost `growth-grown` 26% -- 1.01x to
+        // 1.28x -- for a test whose answer is a precondition here.
+        return a.items[NtsRuntime.bounds(a.length, (int) at)];
     }
 
     public static void set(NtsArrayL a, double at, Object value) {
